more cleaning old pages
