+++ src/app/[locale]/services/page.tsx
@@ -1,10 +1,45 @@
 import type { Metadata } from 'next'
-import ViewServices from './ViewServices'
 
+import { ContactSection } from '@/components/sections/ContactSection'
+import { Container } from '@/components/ui/Container'
+import { PageIntro } from '@/components/sections/PageIntro'
+import { fetchServices, fetchServicesPage } from '@/request/fetch'
+import { ServiceCard } from '@/components/elements/ServiceCard'
 export const metadata: Metadata = {
   title: 'Services',
 }
 
-export default function ViewBlogsPage() {
-  return <ViewServices />
+export default async function ViewServicesPages() {
+  let servicesPageData = null
+  let services = null
+
+  try {
+    servicesPageData = await fetchServicesPage()
+    services = await fetchServices()
+  } catch (error) {
+    console.error('Failed to load data:', error)
+    return <div>Failed to load data</div>
+  }
+
+  const { pageIntro, projectsSection } = servicesPageData || ''
+  // const t = await getTranslations('Project')
+
+  return (
+    <>
+      <PageIntro {...pageIntro}>
+        <p>{pageIntro.content}</p>
+      </PageIntro>
+
+      <Container className="mt-24 sm:mt-32 lg:mt-40">
+        <div className="space-y-24 lg:space-y-32">
+          {services &&
+            services.map((service: any) => (
+              <ServiceCard service={service} key={service.id} />
+            ))}
+        </div>
+      </Container>
+
+      <ContactSection />
+    </>
+  )
 }
